requirements.txt
streamlit
playwright
openai
transformers
torch
requests
Pillow